requirements.txt
beautifulsoup4
selenium
Django==5.1
djangorestframework
markdown
django-filter
psycopg2-binary
django-crontab==0.7.1